+++ iouring_flutter_task/lib/features/watchlist/presentation/bloc/watchlist_bloc.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:iouring_flutter_task/features/watchlist/repository/watchlist_repository.dart';
-import 'package:iouring_flutter_task/features/watchlist/model/watchlist_model.dart';
+import 'package:iouring_flutter_task/features/watchlist/data/repository/watchlist_repository.dart';
+import 'package:iouring_flutter_task/features/watchlist/data/model/watchlist_model.dart';
 import 'watchlist_event.dart';
 import 'watchlist_state.dart';
